Fix 'Failed Unit' error after nodeadm reset (#22)
It turns out that the 1.10.x kubelet binaries exit with non-zero exit codes when terminated.
I’ve tested both 1.10.4 and 1.10.5 and both exhibit this behavior. On the other hand,
something changed in 1.11.0 which fixes this issue.
Terminating kubelet in 1.11.0 produces an exit code of 0.

I've searched through the 1.11 changelogs and I coudn’t find anything mentioning
about an exit code fix for kubelet (there was one for kubectl however).

This exit code issue causes an error to be displayed after login:

    $ ssh root@10.105.16.12
    Last login: Thu Jul 12 18:57:51 UTC 2018 from 10.32.32.139 on pts/0
    Container Linux by CoreOS stable (1745.3.1)
    Update Strategy: No Reboots
    Failed Units: 1
      kubelet.service

Invoking 'systemctl reset-failed kubelet'  fixes the issue.